refactor: remove homepage constants and rename file to toasterConstants
