+++ frontend/tests/constants/toasterConstants.ts
@@ -1,0 +1,13 @@
+import { expect, Page } from "@playwright/test"
+
+export async function assertToastMessage(
+  page: Page,
+  message: string,
+  count: number = 1
+) {
+  await expect(page.locator(".toaster").getByText(message)).toHaveCount(count)
+}
+
+export async function assertToastMessageIsMissing(page: Page, message: string) {
+  await expect(page.locator(".toaster").getByText(message)).toHaveCount(0)
+}
