+++ frontend/tests/homepage.favourite.station.spec.ts
@@ -1,6 +1,6 @@
 import { test } from "./fixture/test"
 import { expect } from "@playwright/test"
-import { assertToastMessageIsMissing } from "./constants/homepageConstants"
+import { assertToastMessageIsMissing } from "./constants/toasterConstants"
 import {
   cantoneseStation,
   stationWithMultipleTags,
