+++ frontend/tests/homepage.share.station.spec.ts
@@ -3,7 +3,7 @@ import { expect } from "@playwright/test"
 import {
   assertToastMessage,
   assertToastMessageIsMissing,
-} from "./constants/homepageConstants"
+} from "./constants/toasterConstants"
 import { stationWithLocationLatLng, unitedStatesStation } from "./mocks/station"
 import { getClipboardContent } from "./constants/shareStationConstants"
 import HomePage from "./pageObjects/HomePage"
