+++ frontend/tests/podcast/detail/podcast.detail.page.share.podcast.spec.ts
@@ -1,7 +1,7 @@
 import { test } from "../../fixture/test"
 import { expect } from "@playwright/test"
 import { defaultTenPodcastEpisodes } from "../../mocks/podcast.episode"
-import { assertToastMessage } from "../../constants/homepageConstants"
+import { assertToastMessage } from "../../constants/toasterConstants"
 import { getClipboardContent } from "../../constants/shareStationConstants"
 import { homePageUrl } from "../../constants/paths"
