+++ frontend/tests/podcast/detail/podcast.detail.page.spec.ts
@@ -1,6 +1,6 @@
 import { test } from "../../fixture/test.ts"
 import { expect } from "@playwright/test"
-import { assertToastMessage } from "../../constants/homepageConstants.ts"
+import { assertToastMessage } from "../../constants/toasterConstants.ts"
 import {
   defaultTenPodcastEpisodes,
   podcastTitleHasPercentSymbol_podcastId_387129_FirstTenEpisodes,
