+++ frontend/tests/podcast/detail/podcast.episode.detail.page.share.spec.ts
@@ -1,7 +1,7 @@
 import { test } from "../../fixture/test"
 import { expect, Page } from "@playwright/test"
 import { podcastId_259760_episodeId_34000697601 } from "../../mocks/podcast.episode"
-import { assertToastMessage } from "../../constants/homepageConstants"
+import { assertToastMessage } from "../../constants/toasterConstants"
 import { getClipboardContent } from "../../constants/shareStationConstants"
 import { homePageUrl } from "../../constants/paths"
